requirements.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
flatbuffers==1.12
gast==0.3.3
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.34
opt-einsum==3.3.0
pillow==10.4.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.6.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.20.2
